integral division result cast to double (did not impact calculations, but added for clarity)

+++ src/net/sourceforge/zmanim/ComplexZmanimCalendar.java
@@ -2772,7 +2772,7 @@ public class ComplexZmanimCalendar extends ZmanimCalendar {
 	 */
 	public Date getFixedLocalChatzos() {
 		return getTimeOffset(getDateFromTime(12.0
-				- getGeoLocation().getTimeZone().getRawOffset() / HOUR_MILLIS),
+				- getGeoLocation().getTimeZone().getRawOffset() / (double)HOUR_MILLIS),
 				-getGeoLocation().getLocalMeanTimeOffset());
 	}
